+++ core/CHANGELOG.md
@@ -1,1 +1,7 @@
 ## HEAD
+
+🤖 PixiJS Engine
+  - 🆕 `components/componentGenerator` v1.0.0.
+  - 🆕 `components/actorGenerator` v1.0.0.
+  - 🆕 `components/functionalities/hitTheWallFunctionalityBuilder` v1.0.0.
+  - 🆕 `components/functionalities/positioningFunctionalityBuilder` v1.0.0.